rename simulated_beacon package to simulatedbeacon.  move under cmd/geth
